refactor(env): PR의 코멘트 내용을 바탕으로 리팩토링 진행 (#7) - App.js 의 필요없는 주석과 fonts.js 파일을 삭제하였다. - devDependencies에 있어야 할 모듈들을 원위치 시켰다. - utils, store 폴더를 만들어 파일 구조들을 정리하였다.
