+++ package.json
@@ -39,6 +39,7 @@
     "babel-jest": "^26.5.2",
     "babel-loader": "^8.1.0",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "css-loader": "^4.3.0",
     "cypress": "^5.3.0",
     "eslint": "^7.10.0",
     "eslint-config-airbnb": "^18.2.0",
@@ -52,8 +53,11 @@
     "html-webpack-plugin": "^4.5.0",
     "jest": "^26.5.2",
     "jest-plugin-context": "^2.9.0",
+    "node-sass": "^4.14.1",
     "react-is": "^16.13.1",
     "redux-mock-store": "^1.5.4",
+    "sass-loader": "^10.0.3",
+    "style-loader": "^2.0.0",
     "url-loader": "^4.1.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12",
@@ -62,16 +66,12 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.4.0",
     "axios": "^0.20.0",
-    "css-loader": "^4.3.0",
-    "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "sass-loader": "^10.0.2",
-    "style-loader": "^1.3.0",
     "styled-components": "^5.2.0"
   }
 }
