--- src/reducer.js
@@ -1,3 +1,0 @@
-import reducer from './slice';
-
-export default reducer;
