+++ src/store/index.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import reducer from './reducer';
+import reducer from './slice';
 
 const store = configureStore({ reducer });
